Change Samples migration to match schema image
This is a tutorial for new people, so it pays to be very consistant
with naming so that new people don't get confused. It is harder
for me to make a new schema image than it is to change the
name of the migration to reflect the name in the schema,
so I changed the migration name of like to liked_the_tea.

On another note, naming is hard.